HacksterBot/requirements.txt
# Core Discord Bot
discord.py>=2.3.2
aiohttp>=3.10.5

# AI Services
pydantic-ai>=0.2.7
openai>=1.3.0
anthropic>=0.34.1

# Database
aiosqlite>=0.19.0

# Search
httpx>=0.27.2

# Content Analysis
beautifulsoup4>=4.12.0
requests>=2.32.3

# Configuration
python-dotenv>=1.0.0

# Logging and Monitoring
colorama>=0.4.6

# File Processing
PyYAML>=6.0.2
emoji>=2.12.1

# Chat Export (for tickets)
chat-exporter>=2.8.0

# URL Processing
pytz>=2024.1
urllib3>=2.2.2

# HTTP Client
aiofiles>=23.2.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
black>=23.7.0
flake8>=6.0.0
mypy>=1.5.0 